# apps/compass-ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Compass AI Backend API",
    "version": "2.1.0",
    "description": "Document analysis with grounded, citation-backed answers, scoped to a client portal. Tenancy is derived from a verified portalFeatureContextToken sent as x-portal-context: NO operation accepts a client, workspace or portal identifier from the caller, and one that carries one is rejected with TENANCY_PROBE. Every operation enforces the derived actor (employee or client) inside its handler. All operations are private: there is no cross-app consumer."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Every failure carries a stable machine code. For a portal failure the message is separately authored to be safe to show a client.",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "TENANCY_PROBE",
                  "PORTAL_NOT_BOUND",
                  "FORBIDDEN_CLIENT_ROLE",
                  "PORTAL_VERIFY_FAILED"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "PortalContext": {
        "name": "x-portal-context",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Platform-signed portalFeatureContextToken from the iframe URL. Verified through Gate on every request; a local decode is never trusted. It proves WHERE the request is from, never WHO — identity comes from the caller session."
      },
...
      "BatchId": {
        "name": "batchId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "TenancyProbe": {
        "description": "The request carried a client/workspace/portal identifier. Refused rather than ignored.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PortalRejected": {
        "description": "Portal context missing, unverifiable, unbound, paused, or the caller is not a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ForbiddenClient": {
        "description": "FORBIDDEN_CLIENT_ROLE — this action is employee-only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found, or it belongs to another client. Deliberately indistinguishable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/indexing": {
      "get": {
        "operationId": "getIndexingSnapshot",
        "summary": "Pipeline health for this client",
        "description": "Employee-only. One batched read, so the whole snapshot comes from a single transaction and cannot contradict itself.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
    "/indexing/batches/{batchId}/cancel": {
      "post": {
        "operationId": "cancelEmbeddingBatch",
        "summary": "Cancel an open embedding batch",
        "description": "Employee-only. Vectors already applied are kept.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/BatchId"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
    "/indexing/batches/{batchId}/poll": {
      "post": {
        "operationId": "pollEmbeddingBatch",
        "summary": "Poll an embedding batch now",
        "description": "Employee-only.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          },
          {
            "$ref": "#/components/parameters/BatchId"
          }
        ],
        "responses": {
          "200": {
            "description": "Outcome"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
    "/indexing/reindex-unsearchable": {
      "post": {
        "operationId": "reindexUnsearchable",
        "summary": "Re-index every unsearchable document",
        "description": "Employee-only.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Enqueued"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
    "/indexing/retry-failed": {
      "post": {
        "operationId": "retryFailedJobs",
        "summary": "Retry failed ingest jobs",
        "description": "Employee-only. Resets attempts and requeues.",
        "x-fusebase-visibility": "private",
        "x-fusebase-execution-mode": "sync",
        "parameters": [
          {
            "$ref": "#/components/parameters/PortalContext"
          }
        ],
        "responses": {
          "200": {
            "description": "Requeued"
          },
          "400": {
            "$ref": "#/components/responses/TenancyProbe"
          },
          "401": {
            "$ref": "#/components/responses/PortalRejected"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenClient"
          },
          "409": {
            "$ref": "#/components/responses/PortalRejected"
          }
        }
      }
    },
...
  }
}